refactor #23

+++ src/HsBundle/Controller/DienstverlenersController.php
@@ -3,13 +3,13 @@
 namespace HsBundle\Controller;
 
 use AppBundle\Controller\SymfonyController;
-use AppBundle\Entity\Klant as AppKlant;
+use AppBundle\Entity\Klant;
 use AppBundle\Form\ConfirmationType;
 use AppBundle\Form\KlantFilterType as AppKlantFilterType;
-use HsBundle\Entity\Klant;
-use HsBundle\Form\KlantFilterType;
-use HsBundle\Form\KlantSelectType;
-use HsBundle\Form\KlantType;
+use AppBundle\Service\KlantDaoInterface;
+use HsBundle\Entity\Dienstverlener;
+use HsBundle\Form\DienstverlenerFilterType;
+use HsBundle\Form\DienstverlenerType;
 use HsBundle\Service\DienstverlenerDaoInterface;
 use JMS\DiExtraBundle\Annotation as DI;
 use Symfony\Component\HttpFoundation\Request;
@@ -29,8 +29,15 @@ class DienstverlenersController extends SymfonyController
      */
     private $dao;
 
+    /**
+     * @var KlantDaoInterface
+     *
+     * @DI\Inject("app.dao.klant")
+     */
+    private $klantDao;
+
     private $enabledFilters = [
-        'openstaand',
+        'rijbewijs',
         'klant' => ['id', 'naam', 'stadsdeel'],
         'filter',
         'download',
@@ -41,17 +48,37 @@ class DienstverlenersController extends SymfonyController
      */
     public function index(Request $request)
     {
-        $filter = $this->getFilter()->handleRequest($request);
-        if ($filter->isSubmitted() && $filter->isValid()) {
-            $pagination = $this->dao->findAll($request->get('page', 1), $filter->getData());
+        $form = $this->getFilter()->handleRequest($request);
+        if ($form->isSubmitted() && $form->isValid()) {
+            $filter = $form->getData();
         } else {
-            $pagination = $this->dao->findAll($request->get('page', 1));
+            $filter = null;
         }
 
+        if ($form->get('download')->isClicked()) {
+            return $this->download($filter);
+        }
+
+        $pagination = $this->dao->findAll($request->get('page', 1), $filter);
+
         return [
-            'filter' => $filter->createView(),
+            'filter' => $form->createView(),
             'pagination' => $pagination,
         ];
+    }
+
+    private function download(FilterInterface $filter)
+    {
+        $collection = $this->dao->findAll(0, $filter);
+
+        $response = $this->render('@Hs/vrijwilligers/download.csv.twig', ['collection' => $collection]);
+
+        $filename = sprintf('homeservice-vrijwilligers-%s.xls', (new \DateTime())->format('d-m-Y'));
+        $response->headers->set('Content-type', 'application/vnd.ms-excel');
+        $response->headers->set('Content-Disposition', sprintf('attachment; filename="%s";', $filename));
+        $response->headers->set('Content-Transfer-Encoding', 'binary');
+
+        return $response;
     }
 
     /**
@@ -61,20 +88,26 @@ class DienstverlenersController extends SymfonyController
     {
         if ($klantId = $request->get('klantId')) {
             if ($klantId === 'new') {
-                $appKlant = new AppKlant();
+                $klant = new Klant();
             } else {
-                $appKlant = $this->getEntityManager()->find(AppKlant::class, $klantId);
+                $klant = $this->klantDao->find($klantId);
             }
 
-            $klant = new Klant($appKlant);
-            $creationForm = $this->getForm($klant)->handleRequest($this->getRequest());
+            // redirect if already exists
+            $dienstverlener = $this->dao->findOneByKlant($klant);
+            if ($dienstverlener) {
+                return $this->redirectToView($dienstverlener);
+            }
+
+            $dienstverlener = new Dienstverlener($klant);
+            $creationForm = $this->getForm($dienstverlener)->handleRequest($this->getRequest());
 
             if ($creationForm->isSubmitted() && $creationForm->isValid()) {
                 try {
-                    $this->dao->create($klant);
-                    $this->addFlash('success', 'Klant is opgeslagen.');
+                    $this->dao->create($dienstverlener);
+                    $this->addFlash('success', 'Dienstverlener is opgeslagen.');
 
-                    return $this->redirectToRoute('hs_klanten_memos_add', ['id' => $klant->getId()]);
+                    return $this->redirectToRoute('hs_dienstverleners_memos_add', ['id' => $dienstverlener->getId()]);
                 } catch (\Exception $e) {
                     $this->addFlash('danger', 'Er is een fout opgetreden.');
                 }
@@ -88,30 +121,21 @@ class DienstverlenersController extends SymfonyController
         }
 
         $filterForm = $this->createForm(AppKlantFilterType::class, null, [
-            'enabled_filters' => ['naam'],
+            'enabled_filters' => ['naam', 'bsn', 'geboortedatum'],
         ]);
         $filterForm->handleRequest($this->getRequest());
 
-        $selectionForm = $this->createForm(KlantSelectType::class, null, [
-            'filter' => $filterForm->getData(),
-        ]);
-        $selectionForm->handleRequest($this->getRequest());
-
         if ($filterForm->isSubmitted() && $filterForm->isValid()) {
+            $builder = $this->getEntityManager()->getRepository(Klant::class)
+                ->createQueryBuilder('klant')
+                ->where('klant.disabled = false')
+                ->orderBy('klant.achternaam')
+            ;
+            $filterForm->getData()->applyTo($builder);
+
             return [
-                'selectionForm' => $selectionForm->createView(),
+                'klanten' => $builder->getQuery()->getResult(),
             ];
-        }
-
-        if ($selectionForm->isSubmitted() && $selectionForm->isValid()) {
-            $klant = $selectionForm->getData();
-            if ($klant->getKlant() instanceof AppKlant) {
-                $klantId = $klant->getKlant()->getId();
-            } else {
-                $klantId = 'new';
-            }
-
-            return $this->redirectToRoute('hs_klanten_add', ['klantId' => $klantId]);
         }
 
         return [
@@ -142,9 +166,9 @@ class DienstverlenersController extends SymfonyController
         if ($form->isSubmitted() && $form->isValid()) {
             try {
                 $this->dao->update($entity);
-                $this->addFlash('success', 'Klant is opgeslagen.');
+                $this->addFlash('success', 'Dienstverlener is opgeslagen.');
 
-                return $this->redirectToView($entity->getId());
+                return $this->redirectToView($entity);
             } catch (\Exception $e) {
                 $this->addFlash('danger', 'Er is een fout opgetreden.');
             }
@@ -166,8 +190,10 @@ class DienstverlenersController extends SymfonyController
         $form = $this->createForm(ConfirmationType::class);
         $form->handleRequest($this->getRequest());
         if ($form->isSubmitted() && $form->isValid()) {
-            $this->dao->delete($entity);
-            $this->addFlash('success', 'Klant is verwijderd.');
+            if ($form->get('yes')->isClicked()) {
+                $this->dao->delete($entity);
+                $this->addFlash('success', 'Dienstverlener is verwijderd.');
+            }
 
             return $this->redirectToIndex();
         }
@@ -180,23 +206,23 @@ class DienstverlenersController extends SymfonyController
 
     private function getFilter()
     {
-        return $this->createForm(KlantFilterType::class, null, [
+        return $this->createForm(DienstverlenerFilterType::class, null, [
             'enabled_filters' => $this->enabledFilters,
         ]);
     }
 
-    private function getForm($data = null)
+    private function getForm(Dienstverlener $data = null)
     {
-        return $this->createForm(KlantType::class, $data);
+        return $this->createForm(DienstverlenerType::class, $data);
     }
 
     private function redirectToIndex()
     {
-        return $this->redirectToRoute('hs_klanten_index');
+        return $this->redirectToRoute('hs_dienstverleners_index');
     }
 
-    private function redirectToView($id)
+    private function redirectToView(Dienstverlener $entity)
     {
-        return $this->redirectToRoute('hs_klanten_view', ['id' => $id]);
+        return $this->redirectToRoute('hs_dienstverleners_view', ['id' => $entity->getId()]);
     }
 }
